fix path to overrides in RN SDK
