+++ packages/sdks/overrides/react-native/src/blocks/text/text.tsx
@@ -1,4 +1,3 @@
-import { registerComponent } from '../functions/register-component';
 import type { BuilderBlock } from '../types/builder-block';
 import HTML from 'react-native-render-html';
 
@@ -66,26 +65,3 @@ export default function Text(props: {
     />
   );
 }
-
-registerComponent(Text, {
-  name: 'Text',
-  static: true,
-  builtIn: true,
-  image:
-    'https://firebasestorage.googleapis.com/v0/b/builder-3b0a2.appspot.com/o/images%2Fbaseline-text_fields-24px%20(1).svg?alt=media&token=12177b73-0ee3-42ca-98c6-0dd003de1929',
-  inputs: [
-    {
-      name: 'text',
-      type: 'html',
-      required: true,
-      autoFocus: true,
-      bubble: true,
-      defaultValue: 'Enter some text...',
-    },
-  ],
-  defaultStyles: {
-    lineHeight: 'normal',
-    height: 'auto',
-    textAlign: 'center',
-  },
-});
